changer commit

+++ tester.py
@@ -1,20 +1,57 @@
 import os
-import sys
+import pathlib
 import datetime
+import pyAesCrypt
+import io
+
+
+def MemoryCrypter(filename, is_encrypted):
+    sequence_bytes = io.BytesIO()
+    bufferSize = 64 * 1024
+    password = "yurayadamvjopueslipoluchitsya"
+
+#True-зашифровать
+#False-расшифровать
+
+    with open(filename, "rb") as f:
+        file_content = io.BytesIO(f.read())
+
+    with open(filename, "wb") as f:
+        if is_encrypted:
+            pyAesCrypt.encryptStream(
+                file_content,
+                sequence_bytes,
+                password,
+                bufferSize
+            )
+        else:
+            pyAesCrypt.decryptStream(
+                file_content,
+                sequence_bytes,
+                password,
+                bufferSize,
+                len(file_content.getvalue())
+            )
+        f.write(sequence_bytes.getvalue())
+
 
 class DEBUG:
     def __init__(self,message):
         level = "DEBUG"
         vremya = str(datetime.datetime.now())
-        stw  ="[" + vremya + ": " + level + "]" +" "+message+"\n"
-        try:
+        stw ="[" + vremya + ": " + level + "]" +" "+message+"\n"
+        path = pathlib.Path('log.txt')
+        if path.exists() == True:
+            MemoryCrypter("log.txt", False)
             my_file = open("log.txt", "a")
             my_file.write(stw)
             my_file.close()
-        except Exception:
+            MemoryCrypter("log.txt", True)
+        else:
             my_file = open("log.txt", "w+")
             my_file.write(stw)
             my_file.close()
+            MemoryCrypter("log.txt",True)
         print(stw)
 
 class INFO:
@@ -22,14 +59,17 @@ class INFO:
         level = "INFO"
         vremya = str(datetime.datetime.now())
         stw = "[" + vremya + ": " + level + "]" +" "+message+"\n"
-        try:
+        path = pathlib.Path('log.txt')
+        if path.exists()==True:
+            MemoryCrypter("log.txt", False)
             my_file = open("log.txt", "a")
             my_file.write(stw)
             my_file.close()
-        except Exception:
+        else:
             my_file = open("log.txt", "w+")
             my_file.write(stw)
             my_file.close()
+            MemoryCrypter("log.txt", True)
         print(stw)
 
 class ERROR:
@@ -37,12 +77,15 @@ class ERROR:
         level = "ERROR"
         vremya = str(datetime.datetime.now())
         stw = "[" + vremya + ": " + level + "]" +" "+message+"\n"
-        try:
+        path = pathlib.Path('log.txt')
+        if path.exists() == True:
+            MemoryCrypter("log.txt", False)
             my_file = open("log.txt", "a")
             my_file.write(stw)
             my_file.close()
-        except Exception:
+        else:
             my_file = open("log.txt", "w+")
             my_file.write(stw)
             my_file.close()
+            MemoryCrypter("log.txt", True)
         print(stw)
